requirements.txt
python-telegram-bot>=21.0
pandas>=2.0.0
watchdog>=4.0.0
python-dotenv>=1.0.0
